removed class and made search a function refactored test

+++ src/main/kotlin/ru/spbstu/ics/Obstacle.kt
@@ -1,4 +1,4 @@
 package ru.spbstu.ics
 
-data class Vertex(val x: Float, val y: Float)
+data class Vertex(val x: Int, val y: Int)
 data class Obstacle(val vertexes: List<Vertex>)
